Preliminary Expressions to Poselibrary feature
- Using directory ops to just read the expressions in Blender.  This
required changes to pass Windows paths with '\'.

- Refactored the base class for syncing ops to not inherit from
bpy.type.Operator.  This is a problem for calling directory ops from
within code.  SyncPose is also called in a loop to transfer all the
expressions (or just those with the tag being filtered by).

- Needed to not call executeJsonCall() in the __init__ of SyncOperator,
so that the filename parameter could be set for SyncPose.

This works in that poses for every expression get created with the
correct name, but the filename being passed of the expression to change
to is not being used yet.  I do not know how to do that.  Anyone know?