avoid crash when adding materials to a mesh without detailed helper vertex-groups
in case of non existent detailed helper vertex-groups the creation of materials
crashed (joint-ground was not there and the poll function crashed with a
context error which will not even fit on the blender page ...)

There are only 3 groups existent, when the detailed groups are not created.
It also would be possible to only color them, but the benefit isn't that
big :)

So I decided to hide the materials checkbox until the detailed helper
is selected. That's easier for the user.

To avoid a crash anyway, we need to check for the detailed
helpers always (other way to crash it: switch on detailed groups, switch on material, switch off
detailed groups ...)